pyproject.toml
[project]
name = "mcp-meetup"
version = "0.1.0"
description = "MCP"
requires-python = ">=3.11.10"
dependencies = [
    "jupyter>=1.1.1",
    "langchain>=0.3.23",
    "langchain-core>=0.3.54",
    "langchain-mcp-adapters>=0.0.9",
    "langchain-openai>=0.3.14",
    "langchain-text-splitters>=0.3.8",
    "langgraph>=0.3.31",
    "mcp[cli]>=1.6.0",
    "openai>=1.75.0",
    "python-dotenv>=1.1.0",
    "yfinance>=0.2.55",
]
